move replacesolution classes to chemistry
didnt make sense for them to be in atmos.rotting
